+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Jamf Plugin for GLPI Changelog
 
-## [1.0.0] unreleased
+## [1.1.0]
+
+### Added
+ - View in Jamf button to mobile device info box in Computers and Phones
+ - When merging items, a GLPI item will be preselected if the UDID/UUID matches (or if the name matches as a backup check)
+ - View and sync Extension Attributes
+
+## [1.0.0]
 
 ### Added
 - Initial release